refactor(form-canvas): move INPUT_CLS to constants.ts to break circular import
controls (validation, options) imported INPUT_CLS from settings-panel, which
in turn imported the controls — a mutual cycle. Extract INPUT_CLS to a leaf
module (constants.ts); all three files now import from there.
