+++ apps/web/src/tools/form-canvas/inspector/constants.ts
@@ -1,0 +1,1 @@
+export const INPUT_CLS = "h-8 w-full rounded-md border border-input bg-background px-2 text-sm";
